fix(frontend): budget split calculation
Previous method of calculating how the monthly total is split is
incorrect.

Only the amount within the budget should be split according to the
percentage allocation. The amount over budget should be split 50/50. The
calculation involving the amount within budget should not include the
over budget amount. Otherwise the split amount will add up to more than
the amount spent.